+++ src/vue-router/index.js
@@ -57,11 +57,6 @@ const router = new Router({
       component: resolve => require(['@/vue/pages/Downloads'], resolve),
     },
     {
-      path: '/ios-installation-guide',
-      name: vueRoutes.iosInstallationGuide.name,
-      component: resolve => require(['@/vue/pages/IosInstallationGuide'], resolve),
-    },
-    {
       path: '/pay',
       name: vueRoutes.pay.name,
       component: resolve => require(['@/vue/pages/Pay'], resolve),
@@ -353,8 +348,14 @@ function authPageGuard (to, from, next) {
 function redirectRouteGuard (to, from, next) {
   const isLoggedIn = store.getters[vuexTypes.isLoggedIn]
   const isAccountUnverified = store.getters[vuexTypes.isAccountUnverified]
+
   if (isLoggedIn && !isAccountUnverified) {
-    if (to.name === vueRoutes.app.name) {
+    const isKycRecoveryInProgress = store
+      .getters[vuexTypes.isKycRecoveryInProgress]
+
+    if (isKycRecoveryInProgress) {
+      next(vueRoutes.kycRecoveryManagement)
+    } else if (to.name === vueRoutes.app.name) {
       const isAccountCorporate = store.getters[vuexTypes.isAccountCorporate]
       const isBusinessToBrowse = store.getters[vuexTypes.isBusinessToBrowse]
       const isCustomerUiShown = store.getters[vuexTypes.isCustomerUiShown]